modifies boards rendering
now, each board is rendered using the same 'createGridLayer' function, with the exception of the board grid, wich besides of using this function also has to render the grid cells.

Also fixes the renderShips function, wich had some minor problems when positioning ships.
And adds the renderAttacks function that displays all the attacks performed in it's own grid layer.