implements ship rotation and improves mouse events
ships can now be rotated by clicking them. Changed the mouse event listeners for pointer event listeners so they also work with other type of pointers.
the adjustSize function will change the size of the ship depending on the rotation.
the getCells function returns an array of the cells the ship occupies by using the row and col value, the length and the rotated state.