changes gameboard's ship cells storing system
ship cells in the gameboard used to have the id as value until they were hit, when they would change to 'true'. This method didn't allow to see from wich ship the cell was after hit, wich would make it very hard to improve the ai logic. Now, all the cells where ships are located have an object value, with two properties: isHit and id. This way, the cells keeps the id of the ship even after a hit was made, wich is indicated by the isHit flag.

Also, the methods isShip and getShipAt were added, for easier manipulation of the new system.