Fixes Ai opponent alogorithm
The ai opponent used to crash when selecting a direction wrong because of multiple ships being next to each other. Now, with the new ship cell system in the gameboard (and the wasAttacked function that was just added), the algorithm has been improved to know when it should change ship direction, and also saves the ships that were accidentaly hit while looking for another ship in a queue, so they can be sunk later.